+++ CHANGELOG.md
@@ -17,12 +17,13 @@
 - `TezosNodeReader.isImplicitAndEmpty` and `TezosNodeReader.isManagerKeyRevealedForAccount` now default to head by reference instead of query.
 - Added `TezosNodeReader.getContractStorage`.
 - Michelson parser support for `D\[UI\]G n`, `D\[UI\]P n`, `DROP n`.
+- Generally improved Michelson contract parser.
 - Improved `TezosContractIntrospector` parser.
 
 ### New Features
 
 - added Tezos Commons Baker Registry interface `chain/tezos/contracts/TCFBakerRegistryHelper`.
-- added TZIP 0007 (fa1.2) token contract interface
+- added TZIP 0007 (FA1.2) token contract interface `chain/tezos/contracts/MorleyTokenHelper`.
 - `TezosMessageUtil` can now `pack` `key_hash` value.
 
 <!-- markdownlint-disable MD024 -->